Add parsing, serialization and module description
Provides parsing and serialization methods for various byte
representations of a Peer ID.

We use the MultihashDigest type from haskell-multihash for the
underlying representation of a PeerId. This representation should make
it easier to ensure that PeerIds are correctly constructed.

Pretty printing of a Peer ID matches the reference implementation at
libp2p/go-libp2p-peer, where the first "Qm" bytes of the multihash are
filtered and six chars of the multihash are pinted.

Also adds a basic module description.