Add quickcheck and unit testing and fixes
Adds quickcheck and unit tests based on the reference tests at
github.com/libp2p/go-libp2p-peer.

Specifically, we add a unit test to check the peerId encoding against
reference keys from the go-libp2p-peer and js-peer-id repos, and
quickcheck tests that ensure a peerid is the same after encoding and
decoding to the PeerId type.

At the moment, we only do testing on RSA Keys. TODO: We need to implement a
better test suite to test over multiple key types.

Dependencies are added and updated as needed to support the changes.